requirements.txt
fastapi==0.111.*
python-dotenv==1.*